Add primitive support for reading streamOptimized VMDK files.
This is not suitable for reading files that are progressively streamed over the network - using the classic GrainDirectory and GrainTable structures to locate grains, rather than the grain markers.

There is also no caching, to avoid redundant de-compression.  This is fine if callers will just sequentially stream out the disk contents in large blocks, but not optimal if accessing a file system within the disk image.